fix(ds): follow-up cleanup — avatar fallback, nanoid dep, drawer rename, toast a11y
Post-merge cleanup of non-blocking items surfaced during the DS-consistency
review:

- Avatar: the sidebar and mobile drawer fed the Avatar a name that already
  defaulted to "Inspector" for nameless users, so avatarInitials produced
  "IN" and the intended fallbackIcon="OI" brand mark never rendered. Pass the
  raw (possibly-empty) name so the "OI" fallback shows; the text label still
  reads "Inspector".
- nanoid: declared as a direct dependency. It was imported at runtime in ~9
  server files but only resolved transitively — a latent break if the
  transitive provider ever drops it.
- InviteSeatModal / FiltersModal -> InviteSeatDrawer / FiltersDrawer: both
  already render the Drawer primitive; the file/component/prop names now match
  what they actually are. Callers, tests, and comments updated.
- Toast: each toast now carries role=alert (errors) / role=status (others)
  plus aria-atomic so screen readers announce them.
