+++ app/components/Toast.tsx
@@ -6,7 +6,7 @@ export function ToastPortal() {
     return (
         <div className="fixed bottom-4 left-1/2 -translate-x-1/2 z-50 flex flex-col gap-2">
             {queue.map(t => (
-                <div key={t.id} className={`bg-ih-bg-card text-ih-fg-1 border border-ih-border rounded-lg shadow-ih-popover px-4 py-2 text-[13px] flex items-center gap-3 min-w-[260px]${
+                <div key={t.id} role={t.variant === 'error' ? 'alert' : 'status'} aria-atomic="true" className={`bg-ih-bg-card text-ih-fg-1 border border-ih-border rounded-lg shadow-ih-popover px-4 py-2 text-[13px] flex items-center gap-3 min-w-[260px]${
                     t.variant === 'error' ? ' border-l-4 border-l-ih-bad'
                     : t.variant === 'warning' ? ' border-l-4 border-l-ih-watch'
                     : t.variant === 'success' ? ' border-l-4 border-l-ih-ok'
