+++ app/components/modals/InviteSeatDrawer.test.tsx
@@ -1,5 +1,5 @@
 /**
- * Free-tier/seat-capped "at limit" gate — the Invite modal should show the
+ * Free-tier/seat-capped "at limit" gate — the Invite drawer should show the
  * seat-limit panel IMMEDIATELY when it opens for a tenant already at its
  * seat cap, instead of only catching the server's 402 SEAT_LIMIT_REACHED
  * after the inviter fills in email/role/permissions and clicks Send invite.
@@ -31,12 +31,12 @@ vi.mock("react-router", async () => {
     };
 });
 
-import { InviteSeatModal } from "~/components/modals/InviteSeatModal";
+import { InviteSeatDrawer } from "~/components/modals/InviteSeatDrawer";
 
-describe("InviteSeatModal — at-open seat-limit gate", () => {
+describe("InviteSeatDrawer — at-open seat-limit gate", () => {
     it("renders the seat-limit panel immediately when seatLimitAtOpen is set (at cap, with billing URL)", () => {
         const { getByText, queryByText } = render(
-            <InviteSeatModal
+            <InviteSeatDrawer
                 open
                 onClose={() => {}}
                 seatLimitAtOpen={{ used: 3, max: 3, billingUrl: "https://billing.example.com" }}
@@ -51,7 +51,7 @@ describe("InviteSeatModal — at-open seat-limit gate", () => {
 
     it("renders the seat-limit panel with no CTA when billingUrl is omitted (no billing portal configured)", () => {
         const { getByText, queryByText } = render(
-            <InviteSeatModal open onClose={() => {}} seatLimitAtOpen={{ used: 3, max: 3 }} />,
+            <InviteSeatDrawer open onClose={() => {}} seatLimitAtOpen={{ used: 3, max: 3 }} />,
         );
         expect(getByText(/Seat limit reached/)).toBeTruthy();
         expect(queryByText("Upgrade")).toBeNull();
@@ -59,7 +59,7 @@ describe("InviteSeatModal — at-open seat-limit gate", () => {
 
     it("renders the normal invite form when under the seat limit (seatLimitAtOpen undefined)", () => {
         const { getByText, queryByText } = render(
-            <InviteSeatModal open onClose={() => {}} seatLimitAtOpen={undefined} />,
+            <InviteSeatDrawer open onClose={() => {}} seatLimitAtOpen={undefined} />,
         );
         expect(queryByText(/Seat limit reached/)).toBeNull();
         expect(getByText("Email")).toBeTruthy();
@@ -67,7 +67,7 @@ describe("InviteSeatModal — at-open seat-limit gate", () => {
     });
 
     it("renders the normal invite form when the prop is omitted entirely (unlimited / other mounts)", () => {
-        const { getByText, queryByText } = render(<InviteSeatModal open onClose={() => {}} />);
+        const { getByText, queryByText } = render(<InviteSeatDrawer open onClose={() => {}} />);
         expect(queryByText(/Seat limit reached/)).toBeNull();
         expect(getByText("Email")).toBeTruthy();
     });
