+++ package.json
@@ -94,6 +94,7 @@
     "hono": "^4.12.25",
     "isbot": "^5.1.0",
     "lib0": "^0.2.117",
+    "nanoid": "^5.1.5",
     "node-diff3": "^3.1.2",
     "otpauth": "^9.5.1",
     "qrcode": "^1.5.4",
